+++ .gitignore
@@ -22,3 +22,6 @@ relation-graph/*
 */.DS_Store
 inferences/*
 node_modules/*
+demo/
+*pyc
+tmp/
