+++ bin/relation-graph
@@ -1,1 +1,1 @@
-
+../relation-graph/bin/relation-graph
