--- environment.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-python3 -m venv venv
-source venv/bin/activate
-export PYTHONPATH=.:$PYTHONPATH
-
