+++ prefixes/prefixes_curated.csv
@@ -8,6 +8,7 @@ dce,http://purl.org/dc/elements/1.1/
 dct,http://purl.org/dc/terms/
 foaf,http://xmlns.com/foaf/0.1/
 protege,http://protege.stanford.edu/plugins/owl/protege#
+skos,http://www.w3.org/2004/02/skos/core#
 ex,http://example.com/
 obo,http://purl.obolibrary.org/obo/
 gocam,http://model.geneontology.org/
@@ -19,3 +20,4 @@ ONTIE,https://ontology.iedb.org/ontology/ONTIE_
 IEDB,http://iedb.org/
 iedb-protein,http://iedb.org/taxon-protein/
 CoVIC,https://cvdb.ontodev.com/antibody/
+EXAMPLE,http://purl.obolibrary.org/obo/EXAMPLE_
