+++ reports/obo.tsv
@@ -1,5 +1,5 @@
-aeo
 agro
+aism
 amphx
 apo
 apollo_sv
@@ -23,6 +23,7 @@ clo
 clyh
 cmo
 cob
+colao
 cro
 cteno
 cto
@@ -30,6 +31,7 @@ cvdo
 ddanat
 ddpheno
 dideo
+disdriv
 doid
 dpo
 dron
@@ -40,6 +42,7 @@ ecocore
 ecto
 emapa
 envo
+epio
 eupath
 exo
 fao
@@ -49,13 +52,12 @@ fbcv
 fbdv
 fideo
 flopo
-fma
 fobi
 foodon
 fovt
 fypo
-gaz
 gecko
+genepio
 geno
 geo
 gno
@@ -75,6 +77,7 @@ ido
 ino
 kisao
 labo
+lepao
 ma
 mamo
 maxo
@@ -100,6 +103,7 @@ nbo
 ncbitaxon
 ncit
 ncro
+nomen
 oae
 oarcs
 oba
@@ -113,12 +117,13 @@ ohd
 ohmi
 ohpi
 olatdv
-omiabis
 omit
 omo
 omp
 omrse
 one
+ons
+ontoavida
 ontoneo
 oostt
 opl
@@ -126,6 +131,7 @@ opmi
 ornaseq
 ovae
 pato
+pcl
 pco
 pdro
 pdumdv
