+++ semsql/ontlib/README.md
@@ -1,0 +1,1 @@
+this sub-package is deprecated and replaced by OAK
