+++ semsql/ontlib/common_queries.py
