--- sql/idmapping.sql
@@ -1,5 +1,0 @@
-CREATE TABLE idmapping (
-  uniprot TEXT,
-  db TEXT,
-  xref TEXT
-);
