+++ tests/test_orm/test_problems.py
@@ -10,8 +10,8 @@ from sqlalchemy import create_engine
 
 
 cwd = os.path.abspath(os.path.dirname(__file__))
-DB_DIR = os.path.join(cwd, 'inputs')
-OUTPUT_DIR = os.path.join(cwd, 'outputs')
+DB_DIR = os.path.join(cwd, '../inputs')
+OUTPUT_DIR = os.path.join(cwd, '../outputs')
 
 class ProblemsTestCase(unittest.TestCase):
